fix warnings on signed/unsigned comparison

+++ toolboxes/mri_core/mri_core_spirit.cpp
@@ -192,21 +192,21 @@ void spirit3d_calib(const hoNDArray<T>& acsSrc, const hoNDArray<T>& acsDst,
         size_t dstCHA = acsDst.get_size(3);
 
         long long kROhalf = kRO / 2;
-        if (2 * kROhalf == kRO)
+        if (2 * static_cast<size_t>(kROhalf) == kRO)
         {
             GWARN_STREAM("spirit3d_calib - 2*kROhalf == kRO " << kRO);
         }
         kRO = 2 * kROhalf + 1;
 
         long long kE1half = kE1 / 2;
-        if (2 * kE1half == kE1)
+        if (2 * static_cast<size_t>(kE1half) == kE1)
         {
             GWARN_STREAM("spirit3d_calib - 2*kE1half == kE1 " << kE1);
         }
         kE1 = 2 * kE1half + 1;
 
         long long kE2half = kE2 / 2;
-        if (2 * kE2half == kE2)
+        if (2 * static_cast<size_t>(kE2half) == kE2)
         {
             GWARN_STREAM("spirit3d_calib - 2*kE2half == kE2 " << kE2);
         }
@@ -217,21 +217,21 @@ void spirit3d_calib(const hoNDArray<T>& acsSrc, const hoNDArray<T>& acsDst,
         if (oE2 > kE2) oE2 = kE2;
 
         long long oROhalf = oRO / 2;
-        if (2 * oROhalf == oRO)
+        if (2 * static_cast<size_t>(oROhalf) == oRO)
         {
             GWARN_STREAM("spirit3d_calib - 2*oROhalf == oRO " << oRO);
         }
         oRO = 2 * oROhalf + 1;
 
         long long oE1half = oE1 / 2;
-        if (2 * oE1half == oE1)
+        if (2 * static_cast<size_t>(oE1half) == oE1)
         {
             GWARN_STREAM("spirit3d_calib - 2*oE1half == oE1 " << oE1);
         }
         oE1 = 2 * oE1half + 1;
 
         long long oE2half = oE2 / 2;
-        if (2 * oE2half == oE2)
+        if (2 * static_cast<size_t>(oE2half) == oE2)
         {
             GWARN_STREAM("spirit3d_calib - 2*oE2half == oE2 " << oE2);
         }
